moved data-retrieval code into functions in searchbox_funcs.py
